% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities_doc.R
\docType{data}
\name{cities}
\alias{cities}
\title{City Names Data}
\format{
\subsection{'cities'}{

A data frame with 4 rows and 3 columns
\describe{
\item{\strong{city}}{Name of city from where the data is gathered}
\item{\strong{lat}}{The latitude value of the city}
\item{\strong{long}}{The longitude value of the city}
}
}
}
\usage{
cities
}
\description{
City Names Data
}
\keyword{datasets}
