% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vehicle_doc.R
\docType{data}
\name{clean_vehicle}
\alias{clean_vehicle}
\title{Electric Vehicle Data}
\format{
\subsection{'clean_vehicle'}{

A data frame with 181038 rows and 13 columns
\describe{
\item{\strong{county}}{Name of county from where the data is gathered}
\item{\strong{city}}{Name of city from where the data is gathered}
\item{\strong{state}}{Name of state from where the data is gathered}
\item{\strong{postal_code}}{The postal code from where the data is present}
\item{\strong{model_year}}{Manufacturing year of the model mentioned in the dataset}
\item{\strong{make}}{Manufacturer of the vehicle}
\item{\strong{model}}{Model Name of the mentioned vehicle}
\item{\strong{electric_vehicle_type}}{Type of the electric vehicle present in the dataset}
\item{\strong{clean_alternative_fuel_vehicle_cafv_eligibility}}{Clean Alternative Fuel Vehicle (CAFV) eligibility for the data present in this dataset}
\item{\strong{electric_range}}{The electric range the car provides}
\item{\strong{vehicle_location}}{Vehicle coordinates}
\item{\strong{lon}}{Longitude of the vehicle's location}
\item{\strong{lat}}{Latitude of the vehicle's location}
}
}
}
\source{
\url{https://www.kaggle.com/datasets/jainaru/electric-vehicle-population/data}
}
\usage{
clean_vehicle
}
\description{
Electric Vehicle Data
}
\keyword{datasets}
