% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-vehicles.R
\name{filter_vehicles}
\alias{filter_vehicles}
\title{Filter Vehicles Based on User Preferences}
\usage{
filter_vehicles(
  data,
  mileage_range = c(0, 500),
  car_make = NULL,
  vehicle_type = NULL
)
}
\arguments{
\item{data}{The dataset to filter (dataframe).}

\item{mileage_range}{A numeric vector of length 2 indicating the range of desired electric mileage.}

\item{car_make}{A character vector of car makes to include.}

\item{vehicle_type}{A character vector of vehicle types to include.}
}
\value{
A filtered dataframe based on the user's provided criteria.
}
\description{
This function filters the dataset based on user-selected electric range,
car make, and type of vehicle.
}
