% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_counts_simplified_doc.R
\docType{data}
\name{grid_counts_simplified}
\alias{grid_counts_simplified}
\title{Simplified Grid Counts Dataset For Leaflet Plot Creation}
\format{
\subsection{'grid_counts_simplified'}{

A data frame with 542 rows and 2 columns
\describe{
\item{\strong{geometry}}{The grouped lat and long values representing a location}
\item{\strong{count}}{The count of EVs in the respective grouped location}
}
}
}
\usage{
grid_counts_simplified
}
\description{
Simplified Grid Counts Dataset For Leaflet Plot Creation
}
\keyword{datasets}
