% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-app.R
\name{load_app}
\alias{load_app}
\title{Loading the Rshiny app}
\usage{
load_app()
}
\description{
This function launches the RShiny app which allows the user to explore the Electric Vehicle
data for themselves and come upon their own conclusions.
}
